// src/models/Drug.ts
/**
 * These are really something like "drug purposes", but they still serve better than magic strings
 */
export enum Drug {
  Antacid = 'antacid',
  AntiBacterial = 'anti bacterial',
  AntiInflammatory = 'anti inflammatory',
  ColdRelief = 'cold relief',
  CoughSuppressant = 'cough suppressant',
  PainRelief = 'pain relief',
}